Start a meager irc client
The architecture is inspired by that of the Tiny client [1]. Currently
it's only possible to connect to one or more servers and exit by
pressing ESC, which is to say, the client is not being a very nice
client.

A debug tab is created when the client is run. Connect to a server by
inputting the command `/connect <hostname>` and pressing ENTER. The
client will try to connect with a hardcoded default nick/user/realname
to the host given on port 6667.

Press TAB to cycle between tabs.

[1] https://github.com/osa1/tiny