Parse PRIVMSG into a more structured type
Make the variant into a struct that contains the type of target, either
a channel or a user, since IRC uses the same PRIVMSG for both query and
channel messages. The `parse_cmd` function receives the parsed vector of
params and can consume from them to populate the more structured ServCmd
variants, but still return the leftovers.