Parse /quit command and goodbye dance
NB. The /quit command is broken due to the simple (and incorrect) input
parsing function. The result of split_whitespace() is checked for a
command first, then everything else is collected into a String, meaning
that all spaces in the message are collapsed. A proper parsing function
will be important.

Once parsed, the /quit command causes a client to send a QUIT message to
its server. According to modern.ircdocs.horse, the server acknowledges
this by replying with an ERROR message and closing the connection.

When the ERROR ServMsg is received, the <reason> argument is printed to
the server tab but the event loop is not broken yet--this only happens
once the lower-level network_loop() receives the Ok(None) from the
TcpStream and sends the Event::Disconnected event up to the event loop.
The network loop breaks after sending the Event::Disconnected message.

The Event::Disconnected message logs a debug line and finally breaks out
of the event loop.

NB. The client is dropped when handle_network_event() exits but this is
merely a clone of the client in the Vec<Client> in the UI. This needs to
be handled separately.

Finally, right now the tabs remain open but since one cannot reconnect
this doesn't make much sense.